Add the option for an admin to delete an artist's profile

+++ src/web/Features/Artists/DeactivateArtist/DeactivateArtistCommand.cs
@@ -4,21 +4,30 @@ using web.Data;
 
 namespace web.Features.Artists.DeactivateArtist;
 
-public class DeactivateArtistCommand(UserManager<IdentityUser<Guid>> userManager,
-        ArtistRepository artistRepository, ApplicationDbContext dbContext)
+public class DeactivateArtistCommand(
+        UserManager<IdentityUser<Guid>> userManager,
+        ArtistRepository artistRepository,
+        ApplicationDbContext dbContext)
 {
-        public async Task ExecuteAsync(Guid userId)
+        public async Task ExecuteAsync(Guid currentUserId, ArtistId artistId)
         {
-                IdentityUser<Guid> user = await userManager.FindByIdAsync(userId.ToString())
-                                              ?? throw new InvalidOperationException("Could not deactivate artist profile - user with such id does not exist.");
-                Artist artist = await artistRepository.GetByUserIdAsync(userId)
+                IdentityUser<Guid> currentUser = await userManager.FindByIdAsync(currentUserId.ToString())
+                                              ?? throw new InvalidOperationException("Could not deactivate artist profile - current user with such id does not exist.");
+                Artist artist = await artistRepository.GetByIdAsync(artistId)
                         ?? throw new InvalidOperationException("No artist exists for a given user.");
+                if (artist.UserId != currentUserId
+                        && !await userManager.IsInRoleAsync(currentUser, Constants.ADMIN_ROLE))
+                {
+                        throw new InvalidOperationException("Attempted to deactivate an artist profile that does not belong to the given user.");
+                }
 
+                IdentityUser<Guid> artistOwner = await userManager.FindByIdAsync(artist.UserId.ToString())
+                                              ?? throw new InvalidOperationException("Could not deactivate artist profile - no user owns this artist's profile.");
                 await dbContext.ArtistPointAwards
                         .Where(award => award.ArtistId == artist.Id)
                         .ExecuteDeleteAsync();
                 artist.Deactivate();
+                await userManager.RemoveFromRoleAsync(artistOwner, Constants.ARTIST_ROLE);
                 await artistRepository.SaveChangesAsync(artist);
-                await userManager.RemoveFromRoleAsync(user, Constants.ARTIST_ROLE);
         }
 }
